requirements.txt
vncorenlp
datasets==1.0.2
transformers
git-python==1.0.3
sacrebleu==1.4.12
rouge_score
PyYAML
torch
pandas
numpy
